Major updates

+++ modules/nf-core/cobrameta/main.nf
@@ -47,12 +47,18 @@ process COBRAMETA {
             --output ${prefix} \\
             $args
 
-        gzip ${prefix}/*.fasta
-        cat ${prefix}/*.fasta.gz > ${prefix}_COBRA_extended.fasta.gz
-        mv ${prefix}/COBRA_joining_summary.txt ${prefix}_COBRA_joining_summary.txt
-        mv ${prefix}/log ${prefix}_log
+        if [ -f ${prefix}/COBRA_extended.fasta ]; then
+            gzip ${prefix}/*.fasta
+            cat ${prefix}/*.fasta.gz > ${prefix}_COBRA_extended.fasta.gz
+            mv ${prefix}/COBRA_joining_summary.txt ${prefix}_COBRA_joining_summary.txt
+            mv ${prefix}/log ${prefix}_log
+        else 
+            cp ${fasta} ${prefix}_COBRA_extended.fasta.gz
+            touch ${prefix}_COBRA_joining_summary.txt
+            mv ${prefix}/log ${prefix}_log
+        fi
     else
-        mv ${fasta} ${prefix}_COBRA_extended.fasta.gz
+        cp ${fasta} ${prefix}_COBRA_extended.fasta.gz
         touch ${prefix}_COBRA_joining_summary.txt
         touch ${prefix}_log
     fi
